requirements.txt
llama-index-llms-groq
llama-index
llama-index-embeddings-huggingface
python-dotenv
streamlit
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.0